Поправил мелкие недоработки
